requirement.txt
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
pip==24.2